+++ CHANGES.rst
@@ -25,6 +25,9 @@ Imviz
 - Footprints plugin for plotting overlays of instrument footprints or custom regions in the image
   viewer. [#2341, #2377]
 
+- Add a curve to stretch histograms in the Plot Options plugin representing the colormap
+  stretch function. [#2390]
+
 Mosviz
 ^^^^^^